Per account nickname
Nickname entry has been added to the account settings. Thus, now it is
possible to have separate nickname on each account. But when it is
empty (initial default) the general setting is used, just as it was
formerly.

This facility is especially useful when participating in several
conferences in different languages where rules require using native
language for nick or it seems merely polite.